+++ src/import.ts
@@ -1,11 +1,11 @@
-import * as fs from 'fs';
-import { v1 as uuidv1 } from 'uuid';
-import * as admin from 'firebase-admin';
+import * as fs from 'fs'
+import { v1 as uuidv1 } from 'uuid'
+import * as admin from 'firebase-admin'
 
 export interface IImportOptions {
-  dates?: string[];
-  geos?: string[];
-  refs?: string[];
+  dates?: string[]
+  geos?: string[]
+  refs?: string[]
 }
 
 /**
@@ -17,11 +17,11 @@ const updateTime = (dateArray, documentObj): void => {
   dateArray.forEach(c => {
     c.split('.').reduce((acc, cur, i, a) => {
       if (!a[i + 1] && acc[cur] && acc[cur]._seconds) {
-        acc[cur] = new Date(acc[cur]._seconds * 1000);
-      } else return acc[cur] || {};
-    }, documentObj);
-  });
-};
+        acc[cur] = new Date(acc[cur]._seconds * 1000)
+      } else return acc[cur] || {}
+    }, documentObj)
+  })
+}
 
 /**
  * Restore data to firestore
@@ -33,91 +33,105 @@ export const restore = (
   fileName: string,
   options: IImportOptions
 ): Promise<any> => {
-  const db = admin.firestore();
+  const db = admin.firestore()
 
   return new Promise((resolve, reject) => {
     if (typeof fileName === 'object') {
-      let dataArray = fileName;
+      let dataObj = fileName
 
-      updateCollection(db, dataArray, options)
+      updateCollection(db, dataObj, options)
         .then(() => {
           resolve({
             status: true,
             message: 'Collection successfully imported!'
-          });
+          })
         })
         .catch(error => {
-          reject({ status: false, message: error.message });
-        });
+          reject({ status: false, message: error.message })
+        })
     } else {
       fs.readFile(fileName, 'utf8', function(err, data) {
         if (err) {
-          console.log(err);
-          reject({ status: false, message: err.message });
+          console.log(err)
+          reject({ status: false, message: err.message })
         }
 
         // Turn string from file to an Array
-        let dataArray = JSON.parse(data);
+        let dataObj = JSON.parse(data)
 
-        updateCollection(db, dataArray, options)
+        updateCollection(db, dataObj, options)
           .then(() => {
             resolve({
               status: true,
               message: 'Collection successfully imported!'
-            });
+            })
           })
           .catch(error => {
-            reject({ status: false, message: error.message });
-          });
-      });
+            reject({ status: false, message: error.message })
+          })
+      })
     }
-  }).catch(error => console.error(error));
-};
+  }).catch(error => console.error(error))
+}
 
 /**
  * Update data to firestore
  *
  * @param {any} db
- * @param {Array<any>} dataArray
- * @param {Array<string>} dateArray
- * @param {Array<string>} geoArray
+ * @param {object} dataObj
+ * @param {IImportOptions} options
  */
 const updateCollection = async (
   db,
-  dataArray: Array<any>,
+  dataObj: object,
   options: IImportOptions = {}
 ) => {
-  for (var index in dataArray) {
-    var collectionName = index;
-    for (var doc in dataArray[index]) {
-      if (dataArray[index].hasOwnProperty(doc)) {
+  for (var index in dataObj) {
+    var collectionName = index
+    for (var doc in dataObj[index]) {
+      if (dataObj[index].hasOwnProperty(doc)) {
         // asign document id for array type
-        let docId = Array.isArray(dataArray[index]) ? uuidv1() : doc;
-        if (dataArray[index][doc]['subCollection']) {
-          const subCollections = dataArray[index][docId]['subCollection'];
-          delete dataArray[index][doc]['subCollection'];
+        let docId = Array.isArray(dataObj[index]) ? uuidv1() : doc
+        if (!Array.isArray(dataObj[index])) {
+          const subCollections = dataObj[index][docId]['subCollection']
+          delete dataObj[index][doc]['subCollection']
           await startUpdating(
             db,
             collectionName,
             docId,
-            dataArray[index][doc],
+            dataObj[index][doc],
             options
-          );
-          // Update sub collection
-          await updateCollection(db, subCollections);
+          )
+
+          if (subCollections) {
+            await updateCollection(db, subCollections, options)
+          }
         } else {
+          const subCollections = dataObj[index][doc]['subCollection']
+
+          delete dataObj[index][doc]['subCollection']
+
           await startUpdating(
             db,
             collectionName,
             docId,
-            dataArray[index][doc],
+            dataObj[index][doc],
             options
-          );
+          )
+
+          if (subCollections) {
+            for (const subIndex in subCollections) {
+              const revivedSubCollection = {}
+              const subCollectionPath = `${collectionName}/${docId}/${subIndex}`
+              revivedSubCollection[subCollectionPath] = subCollections[subIndex]
+              await updateCollection(db, revivedSubCollection, options)
+            }
+          }
         }
       }
     }
   }
-};
+}
 
 /**
  * Write data to database
@@ -125,8 +139,7 @@ const updateCollection = async (
  * @param collectionName
  * @param docId
  * @param data
- * @param dateArray
- * @param geoArray
+ * @param options
  */
 
 const startUpdating = (
@@ -138,16 +151,21 @@ const startUpdating = (
 ) => {
   // Update date value
   if (options.dates && options.dates.length > 0) {
-    updateTime(options.dates, data);
+    updateTime(options.dates, data)
   }
 
   // reference key
   if (options.refs && options.refs.length > 0) {
     options.refs.forEach(ref => {
       if (data.hasOwnProperty(ref)) {
-        data[ref] = db.doc(data[ref]);
+        // check type of the reference
+        if (Array.isArray(data[ref])) {
+          data[ref] = data[ref].map(ref => db.doc(ref))
+        } else {
+          data[ref] = db.doc(data[ref])
+        }
       }
-    });
+    })
   }
 
   // Enter geo value
@@ -157,11 +175,11 @@ const startUpdating = (
         data[geo] = new admin.firestore.GeoPoint(
           data[geo]._latitude,
           data[geo]._longitude
-        );
+        )
       } else {
-        console.warn('Please check your geo parameters!!!', options.geos);
+        console.warn('Please check your geo parameters!!!', options.geos)
       }
-    });
+    })
   }
 
   return new Promise((resolve, reject) => {
@@ -169,18 +187,18 @@ const startUpdating = (
       .doc(docId)
       .set(data)
       .then(() => {
-        console.log(`${docId} was successfully added to firestore!`);
+        console.log(`${docId} was successfully added to firestore!`)
         resolve({
           status: true,
           message: `${docId} was successfully added to firestore!`
-        });
+        })
       })
       .catch(error => {
-        console.log(error);
+        console.log(error)
         reject({
           status: false,
           message: error.message
-        });
-      });
-  });
-};
+        })
+      })
+  })
+}